Implement Drafts with simplified versioning concept (#835)
* Exclude sent templates from Drafts

* Sort sent messages reverse-chronologically

* change some labels

* Hide show all link if under threshold

* Hide previous versions for now

* Just save over drafts (remove versioning)

* Remove draft versioning, add latest option flag (no longer needed but leaving anyway)

* fix lint errors

* Add messages when there are no Drafts or Sent

* Return to messages after save

* No more redirects to base

* Add confirmation dialog on delete template

* Update deps

* Add more test coverage

* Add MessageFactory tests (and fix update_object)

* More test coverage

* More test coverage

* Fix potentially flaky test

* Only display Messages title on main messages screen

* Rename route to drafts

* Add back to messages links

* Add more back links

* Update deps

* try force resolve deps

* revert package-lock to last known working

* revert further

* Try updating cypress action

* Add force flag

+++ wordpress/wp-content/plugins/gc-lists/resources/js/src/messages/components/ListTemplates.tsx
@@ -8,6 +8,7 @@ import { Spinner } from "../../common/Spinner";
 import { Table, StyledLink, StyledPaging } from "./Table";
 import { Next } from "./icons/Next";
 import useTemplateApi from '../../store/useTemplateApi';
+import { ConfirmDelete } from './ConfirmDelete';
 
 const StyledDivider = styled.span`
     margin-left: 10px;
@@ -85,8 +86,11 @@ export const ListTemplates = ({ perPage, pageNav }: { perPage?: number, pageNav?
                             <StyledDivider>|</StyledDivider>
                             <StyledDeleteButton data-tid={tId}
                                 onClick={async () => {
-                                    await deleteTemplate({ templateId: tId });
-                                    getTemplates();
+                                    const confirmed = await ConfirmDelete();
+                                    if (confirmed) {
+                                        await deleteTemplate({ templateId: tId });
+                                        getTemplates();
+                                    }
                                 }}
                             >
                                 {__("Delete", "gc-lists")}
@@ -107,26 +111,37 @@ export const ListTemplates = ({ perPage, pageNav }: { perPage?: number, pageNav?
 
     return (
         <>
+            {window.location.hash === '#/messages' &&
+                <h1>{__('Messages', 'gc-lists')}</h1>
+            }
             <Link
                 className="button button-primary"
                 to={`/messages/edit/new`}
             >
                 {__("Create new message", "gc-lists")}
             </Link>
-
+            <h2>{__('Draft messages', 'gc-lists')}</h2>
             {loading && <Spinner />}
             {
                 templates?.length ?
                     <>
-                        <h2>{__('Saved messages', 'gc-lists')}</h2>
                         <Table columns={columns} data={templates} perPage={perPage} pageNav={pageNav} />
-                        <StyledPaging>
-                            <StyledLink to={`/messages/all-templates`} >
-                                <span>{__("All saved message", "gc-lists")}</span><Next />
-                            </StyledLink>
-                        </StyledPaging>
+                        {templates?.length > 6 &&
+                            <StyledPaging>
+                                <StyledLink to={`/messages/all-drafts`}>
+                                    <span>{__("All drafts", "gc-lists")}</span><Next />
+                                </StyledLink>
+                            </StyledPaging>
+                        }
                     </>
-                    : null
+                    :
+                    <>
+                        {!loading &&
+                            <>
+                                <p>{__("You have no draft messages", "gc-lists")}</p>
+                            </>
+                        }
+                    </>
             }
         </>
     )
